+++ Java-veebirakendus/src/main/java/ee/itcollege/group04_2011/entities/Seadus.java
@@ -30,7 +30,7 @@ public class Seadus implements Serializable {
 	@Id
 	@GeneratedValue(strategy=GenerationType.AUTO)
 	@Column(name="SEADUSE_ID")
-	private int seaduseId;
+	private Long seaduseId;
 
 	private String avaja;
 
@@ -66,11 +66,11 @@ public class Seadus implements Serializable {
     public Seadus() {
     }
 
-	public int getSeaduseId() {
+	public Long getSeaduseId() {
 		return this.seaduseId;
 	}
 
-	public void setSeaduseId(int seaduseId) {
+	public void setSeaduseId(Long seaduseId) {
 		this.seaduseId = seaduseId;
 	}
